feature(order-product): Implementar insercao e delete

+++ src/modules/app/services/ProductService.ts
@@ -1,5 +1,5 @@
-import { Injectable } from '@nestjs/common';
-import { ProductRepository } from 'modules/app/repositories/product';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { ProductRepository } from 'modules/app/repositories/ProductRepository';
 import { IProduct } from 'modules/database/interfaces/IProduct';
 
 @Injectable()
@@ -7,7 +7,10 @@ export class ProductService {
   constructor(private productRepository: ProductRepository) {}
 
   public async remove(id: number) {
-    this.productRepository.remove(id);
+    const product = this.productRepository.findById(id);
+    if (!product) throw new NotFoundException('not-found');
+
+    return this.productRepository.remove(id);
   }
 
   public async save(model: IProduct) {
